backend/requirements.txt
# Core LLM — NVIDIA NIM via OpenAI-compatible client
openai>=1.30.0
python-dotenv>=1.0.0

# PDF processing
pymupdf>=1.24.0
pymupdf4llm>=0.0.12
Pillow>=10.0.0

# Notebook creation
nbformat>=5.10.0

# ── NVIDIA Agentic AI ──
nvidia-nat[langchain,phoenix]>=1.3.0
langchain>=0.3.0
langchain-nvidia-ai-endpoints>=0.3.0
langgraph>=0.2.0

# ── Observability — Arize Phoenix ──
arize-phoenix>=8.0.0
arize-phoenix-otel>=0.8.0
openinference-instrumentation-openai>=0.1.0

# ── Web Search & Repo Discovery (replaces ScrapingBee) ──
valyu>=2.5.0
requests>=2.31.0

# Utilities
rich>=13.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
tiktoken>=0.7.0

# ── NVIDIA NeMo Data Designer (OPTIONAL — install separately if needed) ──
# pip install data-designer
# Not included here because it causes pip resolution-too-deep errors
# with arize-phoenix. The synthetic_data.py module has inline fallbacks
# that work without this package.